Moved code from module "mongock-spring-runner-api" to "mongock-springbook-base". Deleted module "mongock-spring-runner-api" (unnecessary). Generalized driver context selector (dynamic).
