+++ mongock-core-bom/pom.xml
@@ -52,11 +52,6 @@
       </dependency>
       <dependency>
         <groupId>com.github.cloudyrock.mongock</groupId>
-        <artifactId>mongock-spring-runner-api</artifactId>
-        <version>5.0.3.BETA-SNAPSHOT</version>
-      </dependency>
-      <dependency>
-        <groupId>com.github.cloudyrock.mongock</groupId>
         <artifactId>mongock-springboot-base</artifactId>
         <version>5.0.3.BETA-SNAPSHOT</version>
       </dependency>
