+++ mongock-runner/spring/pom.xml
@@ -13,7 +13,6 @@
   <name>Mongock Spring 5 runner</name>
   <description>Spring runner for Mongock</description>
   <modules>
-    <module>mongock-spring-runner-api</module>
     <module>springboot</module>
   </modules>
