+++ mongock-runner/spring/springboot/mongock-springboot-base/src/main/java/com/github/cloudyrock/springboot/base/config/SpringRunnerType.java
@@ -1,4 +1,4 @@
-package com.github.cloudyrock.spring.config;
+package com.github.cloudyrock.springboot.base.config;
 
 public enum SpringRunnerType {
   InitializingBean, ApplicationRunner
