+++ mongock-runner/spring/springboot/mongock-springboot-base/src/main/java/com/github/cloudyrock/springboot/base/util/ProfileUtil.java
@@ -1,4 +1,4 @@
-package com.github.cloudyrock.spring.util;
+package com.github.cloudyrock.springboot.base.util;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.AnnotatedElement;
