+++ mongock-runner/spring/springboot/mongock-springboot-base/src/main/java/com/github/cloudyrock/springboot/base/util/importers/ArtifactDescriptor.java
@@ -1,4 +1,4 @@
-package com.github.cloudyrock.spring.util.importers;
+package com.github.cloudyrock.springboot.base.util.importers;
 
 public class ArtifactDescriptor {
