+++ mongock-runner/spring/springboot/mongock-springboot-base/src/main/java/com/github/cloudyrock/springboot/base/util/importers/MongockDriverContextSelectorUtil.java
@@ -1,8 +1,7 @@
-package com.github.cloudyrock.spring.util.importers;
+package com.github.cloudyrock.springboot.base.util.importers;
 
 import com.github.cloudyrock.mongock.exception.MongockException;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -11,29 +10,24 @@ public class MongockDriverContextSelectorUtil {
   private final static String DRIVER_NOT_FOUND_ERROR_TEMPLATE = "MONGOCK DRIVER HAS NOT BEEN IMPORTED" +
       "\n====================================" +
       "\n\tSOLUTION: You need to import one of the following artifacts";
-
-  private static final List<ContextImporter> contextImporters = Collections.singletonList(
-      new MongoSpringDataImporter()
-  );
-
-  private static final String DRIVER_NOT_FOUND_ERROR;
-
-  static {
+  
+  private static String generateDriverNotFoundError (List<ContextImporter> contextImporters) {
     StringBuilder sb = new StringBuilder(DRIVER_NOT_FOUND_ERROR_TEMPLATE);
     contextImporters.stream()
         .map(ContextImporter::getArtifacts)
         .flatMap(List::stream)
         .forEach(desc -> sb.append("\n\t- '").append(desc.getArtifact()).append("' for ").append(desc.getTitle()));
-    DRIVER_NOT_FOUND_ERROR = sb.toString();
+    return sb.toString();
   }
-
-
-  public static String[] selectImports() {
+  
+  public static String[] selectImports(List<ContextImporter> contextImporters) {
     return contextImporters.stream()
         .map(contextImporter -> contextImporter.getPaths())
         .filter(Objects::nonNull)
         .findFirst()
-        .orElseThrow(() -> new MongockException(String.format("\n\n%s\n\n", DRIVER_NOT_FOUND_ERROR)));
+        .orElseThrow(() -> new MongockException(String.format("\n\n%s\n\n", generateDriverNotFoundError(contextImporters))));
   }
+  
+
 
 }
