+++ mongock-runner/spring/springboot/mongock-springboot/src/main/java/com/github/cloudyrock/springboot/config/importers/MongoSpringDataImporter.java
@@ -1,5 +1,7 @@
-package com.github.cloudyrock.spring.util.importers;
+package com.github.cloudyrock.springboot.config.importers;
 
+import com.github.cloudyrock.springboot.base.util.importers.ArtifactDescriptor;
+import com.github.cloudyrock.springboot.base.util.importers.ContextImporter;
 import java.util.Arrays;
 import java.util.List;
 
@@ -8,6 +10,8 @@ public class MongoSpringDataImporter implements ContextImporter {
   private final static String PACKAGE_TEMPLATE = "com.github.cloudyrock.mongock.driver.mongodb.springdata.v%s.";
   private final static String DRIVER_TEMPLATE = PACKAGE_TEMPLATE + "SpringDataMongoV%sDriver";
   private final static String CONTEXT_TEMPLATE = PACKAGE_TEMPLATE + "config.SpringDataMongoV%sContext";
+  private final static String ARTIFACT_TITLE_TEMPLATE = "MongoDB Spring data %s";
+  private final static String ARTIFACT_IDENTIFIER_TEMPLATE = "com.github.cloudyrock.mongock:mongodb-springdata-v%s-driver";
 
   @Override
   public String[] getPaths() {
@@ -31,7 +35,7 @@ public class MongoSpringDataImporter implements ContextImporter {
   }
 
   private ArtifactDescriptor getArtifactDescriptor(String v) {
-    return new ArtifactDescriptor("MongoDB Spring data " + v, "com.github.cloudyrock.mongock:mongodb-springdata-v" + v + "-driver");
+    return new ArtifactDescriptor(String.format(ARTIFACT_TITLE_TEMPLATE, v), String.format(ARTIFACT_IDENTIFIER_TEMPLATE, v));
   }
 
   private String[] loadSpringDataContextByVersion(String v) throws ClassNotFoundException {
